refactor(constants): move to constants/form, constants/messages
