+++ constants/messages.ts
@@ -1,0 +1,6 @@
+const TOAST_MESSAGES = {
+  SURVEY_SUBMITTED: '설문이 제출되었습니다. 감사합니다!',
+  SURVEY_SUBMIT_FAILED: '설문 제출에 실패했습니다. 다시 시도해주세요.',
+} as const;
+
+export { TOAST_MESSAGES };
