+++ lib/api-instance.ts
@@ -1,4 +1,5 @@
 import { Forms } from '@/__generated__/Forms';
+import { type CreateFormDto } from '@/__generated__/data-contracts';
 import { HTTP_HEADERS } from '@/constants/http-headers';
 
 const formsApiInstance = new Forms({
@@ -8,4 +9,15 @@ const formsApiInstance = new Forms({
   },
 });
 
-export { formsApiInstance };
+const axiosPostForm = async (createFormDto: CreateFormDto) => {
+  if (!process.env.NEXT_PUBLIC_SCHEMA_SLUG) {
+    throw new Error('NEXT_PUBLIC_SCHEMA_SLUG is not defined');
+  }
+
+  await formsApiInstance.formsControllerCreate(
+    process.env.NEXT_PUBLIC_SCHEMA_SLUG,
+    createFormDto,
+  );
+};
+
+export { axiosPostForm, formsApiInstance };
